requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# OCR
paddleocr>=2.6.0
paddlepaddle>=2.4.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
